requirements.txt
flask
flask-smorest
gunicorn
flask-cors
flask-sqlalchemy
marshmallow
pg8000
google-cloud-secret-manager